+++ core/web/jsonld/EventTrait.php
@@ -78,7 +78,7 @@ trait EventTrait
      */
     public function setAttendee($attendee)
     {
-        ObjectHelper::instanceOf($attendee, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($attendee, [Organization::class, PersonInterface::class]);
         
         $this->_attendee = $attendee;
         return $this;
@@ -102,7 +102,7 @@ trait EventTrait
      */
     public function setComposer($composer)
     {
-        ObjectHelper::instanceOf($author, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($author, [Organization::class, PersonInterface::class]);
         
         $this->_composer = $composer;
         return $this;
@@ -126,7 +126,7 @@ trait EventTrait
      */
     public function setContributor($contributor)
     {
-        ObjectHelper::instanceOf($contributor, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($contributor, [Organization::class, PersonInterface::class]);
         
         $this->_contributor = $contributor;
         return $this;
@@ -240,7 +240,7 @@ trait EventTrait
      */
     public function setFunder($funder)
     {
-        ObjectHelper::instanceOf($funder, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($funder, [Organization::class, PersonInterface::class]);
         
         $this->_funder = $funder;
         return $this;
@@ -353,7 +353,7 @@ trait EventTrait
      */
     public function setOrganizer($organizer)
     {
-        ObjectHelper::instanceOf($organizer, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($organizer, [Organization::class, PersonInterface::class]);
         
         $this->_organizer = $organizer;
         return $this;
@@ -378,7 +378,7 @@ trait EventTrait
      */
     public function setPerformer($performer)
     {
-        ObjectHelper::instanceOf($performer, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($performer, [Organization::class, PersonInterface::class]);
         
         $this->_performer = $performer;
         return $this;
@@ -474,7 +474,7 @@ trait EventTrait
      */
     public function setSponsor($sponsor)
     {
-        ObjectHelper::instanceOf($sponsor, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($sponsor, [Organization::class, PersonInterface::class]);
         
         $this->_sponsor = $sponsor;
         return $this;
@@ -568,7 +568,7 @@ trait EventTrait
      */
     public function setTranslator($translator)
     {
-        ObjectHelper::instanceOf($translator, [Organization::class, PersonInterface::class]);
+        ObjectHelper::isInstanceOf($translator, [Organization::class, PersonInterface::class]);
         
         $this->_translator = $translator;
         return $this;